refactor: 인젝터 에러문구 추가 및 코드 리팩터링
