+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import woowacourse.shopping.di.sgDi.annotation.WooWaField
+import woowacourse.shopping.di.sgDi.annotation.WooWaQualifier
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.repository.CartRepository
-import woowacourse.shopping.util.annotation.WooWaField
-import woowacourse.shopping.util.annotation.WooWaQualifier
 
 class CartViewModel : ViewModel() {
