+++ app/src/main/java/woowacourse/shopping/ui/main/MainViewModel.kt
@@ -5,13 +5,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import woowacourse.shopping.di.sgDi.annotation.WooWaQualifier
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
-import woowacourse.shopping.util.annotation.WooWaInject
-import woowacourse.shopping.util.annotation.WooWaQualifier
 
-class MainViewModel @WooWaInject constructor(
+class MainViewModel(
     private val productRepository: ProductRepository,
     @WooWaQualifier(WooWaQualifier.DATABASE)
     private val cartRepository: CartRepository,
